+++ CHANGELOG.md
@@ -8,7 +8,9 @@ The project uses semantic versioning (see [semver](https://semver.org)).
 
 ### Added
 - Forges now discovered via setuptools
-- New GithubForge implemented
+- New "run-forge" subcommand for creating PRs for on-disk branches
+- Implemented GithubForge for PR creation on Github
+- New "Stamper" `PatchDriver` for "stamping" new files on repos
 
 ### Fixed
 - Incorrect Repo creation causing silent cloning errors on some devices